Modified dunst template and enhanced wal reload.
Added current time when starting pomodoro script.

+++ home/.config/wal/reload_services.sh
@@ -1,4 +1,12 @@
 #!/usr/bin/env bash
 
+prog_exists() {
+    printf "$(command -v >/dev/null 2>&1)"
+}
+
 # Dunst
-killall dunst
+if prog_exists "dunst"; then
+    killall dunst
+fi
+
+notify-send "pywal" "Reloaded wal configurations!"
